requirements.txt
python-dotenv
flask
flask_socketio
flask_cors
python-dateutil
openai
requests
google-generativeai
opencc-python-reimplemented
pandas
pdfplumber
frontend
sqlalchemy
psycopg2
PyMuPDF==1.23.8
tools
openpyxl